Bugfix: blended output image has transparent areas.
Before, when blending the source image and the sorted image, the
transparency of the source image would be blended aswell, giving the
output image some transparent areas.
However, the transparency of the input is only indended to be used as a
mask for pixel sorting, and the output should be fully opaque.

Now, the source and sorted images are converted to RGB - which has no
alpha channel - before they are blended.